Improved cross-python compatibility and added Github Actions matrix strategy
The setup.py script has been updated to ensure 'virtualenv' is only required for Python 2. Additionally, this script now more accurately specifies the versions of Python that are supported. In addition to these changes, the Github Actions workflow file is updated to use a matrix strategy to test against multiple versions of Python. Lastly, the project version number is updated and various functional improvements made in the codebase for better maintenability.